Moved ghost node tests to the navigation module
